Add launch time
