+++ starport/services/network/networktypes/chainlaunch_test.go
@@ -34,18 +34,20 @@ func TestParseChainLaunch(t *testing.T) {
 				ChainID:     "foo-1",
 				SourceURL:   "foo.com",
 				SourceHash:  "0xaaa",
-				CampaignID:  1,
 				GenesisURL:  "",
 				GenesisHash: "",
+				CampaignID:  1,
 			},
 		},
 		{
-			name: "chain with custom genesis url and no campaign",
+			name: "launched chain with custom genesis url and no campaign",
 			fetched: launchtypes.Chain{
 				LaunchID:       1,
 				GenesisChainID: "bar-1",
 				SourceURL:      "bar.com",
 				SourceHash:     "0xbbb",
+				LaunchTriggered: true,
+				LaunchTimestamp: 100,
 				InitialGenesis: launchtypes.NewGenesisURL(
 					"genesisfoo.com",
 					"0xccc",
@@ -56,9 +58,10 @@ func TestParseChainLaunch(t *testing.T) {
 				ChainID:     "bar-1",
 				SourceURL:   "bar.com",
 				SourceHash:  "0xbbb",
-				CampaignID:  0,
 				GenesisURL:  "genesisfoo.com",
 				GenesisHash: "0xccc",
+				LaunchTime: 100,
+				CampaignID:  0,
 			},
 		},
 	}
